Fixes a bug with the configuration handling.
Previously, we were overwriting property configuration data with
the `tasty-hedgehog` defaults.

Now, we don't change these values if we can detect that the user
has changed them, and we don't mess with the defaults unless
the user has explicitly passed something in via the command line.

Also removes the verbosity flag, which was having no effect.
I'm not sure if we can easily add that back in, although I'll look
into it later on.